refactor: put scripts in directories
