+++ Bug_Rider/Assets/Scripts/Respawn/RespawnTriggerPoint.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+public class RespawnTriggerPoint : MonoBehaviour
+{
+    public RespawnManager manager;
+    public int index;
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            manager.SetRespawnIndex(index);
+        }
+    }
+}
